updates tests to make an explicitly fake django-page
the homepage require a bucket on the server to render, which would
require a `server.create('bucket')` call in every acceptance test's
beforeEach hook. OR we could create a fake django-page which already has
its server dependencies baked into its create process.

I went for the latter b/c it requires less boilerplate and one less
thing to remember, in a sense. we still need to remember to make
a django-page that isn't the homepage, but at least the pattern exists
closer to the test at hand, i.e. we don't need to remmber to look in the
beforeEAch for a clue.