make listing pages conform to story slugs as ids
stories are configured to be looked up using the value of the slug
attribute. this patch updates how api-response payloads are serialized.
the sideloaded records need their ids to be set to the value of the that
record's slug.

why do this on the client and not the server? it's my opinion that we
should keep IDs as the actual database ID in the serialized output from
the endpoints; our client uses slugs to navigate, but making the id
a slug for all clients seems dangerous.