change logic to determine stream vs on demand
now that we use slugs to ID stories, the regex wasn't valid. this change
uses an array of known stream names to determine if a given slug is for
a stream or not.

once audio services lands, this file will be gone and we'll be able to
attach ember data records to play actions, which can then determine how
to be played by dj/hifi