include ember-cli history middleware to enable fastboot
long story short, the `history-support` middleware needs to be enabled in order for `ember-cli-fastboot` to handle requests to routes.

the reason we need this is [this line](https://github.com/ember-cli/ember-cli/blob/6151dec2791d255751f3188e14c81c4fa2a1dc5f/lib/tasks/server/middleware/history-support/index.js#L30), which will only include the `history-support` middleware if the `locationType` is one of `auto` or `history`. the wqxr app has been configured to use a custom location type named `trailing-history`. fortunately `history-support` allows for a specific config to turn it on, which is what this commit sets.